+++ drivers/media/pci/ddbridge/ddbridge-core.c
@@ -2705,7 +2705,7 @@ static long ddb_ioctl(struct file *file, unsigned int cmd, user_uintptr_t arg)
 	struct ddb *dev = file->private_data;
 
 	dev_warn(dev->dev, "DDB IOCTLs unsupported (cmd: %d, arg: %lu)\n",
-		 cmd, arg);
+		 cmd, (unsigned long)arg);
 
 	return -ENOTTY;
 }
