+++ drivers/tty/serial/8250/8250_pci.c
@@ -4302,7 +4302,7 @@ pciserial_init_one(struct pci_dev *dev, const struct pci_device_id *ent)
 	}
 
 	if (ent->driver_data >= ARRAY_SIZE(pci_boards)) {
-		pci_err(dev, "invalid driver_data: %ld\n", ent->driver_data);
+		pci_err(dev, "invalid driver_data: %ld\n", (unsigned long)ent->driver_data);
 		return -EINVAL;
 	}
